+++ compiler-cli/src/dependencies.rs
@@ -635,7 +635,7 @@ fn resolve_versions<Telem: Telemetry>(
     // The version requires of the current project
     let mut root_requirements = HashMap::new();
 
-    // Populate the provided_packages and root_requrements maps
+    // Populate the provided_packages and root_requirements maps
     for (name, requirement) in dependencies.into_iter() {
         let version = match requirement {
             Requirement::Hex { version } => version,
@@ -668,7 +668,7 @@ fn resolve_versions<Telem: Telemetry>(
         &locked,
     )?;
 
-    // Convert the hex packages and local packages into manfiest packages
+    // Convert the hex packages and local packages into manliest packages
     let manifest_packages = runtime.block_on(future::try_join_all(
         resolved
             .into_iter()
@@ -721,7 +721,7 @@ fn provide_git_package(
         repo: SmolStr::new_inline("repo"),
         commit: SmolStr::new_inline("commit"),
     };
-    Err(Error::GitDependencyUnsuported)
+    Err(Error::GitDependencyUnsupported)
 }
 
 /// Adds a gleam project located at a specific path to the list of "provided packages"
@@ -733,7 +733,7 @@ fn provide_package(
     provided: &mut HashMap<SmolStr, ProvidedPackage>,
     parents: &mut Vec<SmolStr>,
 ) -> Result<hexpm::version::Range> {
-    // Return early if a package cyle is detected
+    // Return early if a package cycle is detected
     if parents.contains(&package_name) {
         let mut last_cycle = parents
             .split(|p| p == &package_name)