+++ compiler-core/src/javascript/tests/bit_strings.rs
@@ -242,3 +242,24 @@ fn go(x) {
 "#,
     );
 }
+
+#[test]
+fn as_module_const() {
+    assert_js!(
+        r#"
+          pub const data = <<
+            0x1,
+            2,
+            2:size(16),
+            0x4:size(32),
+            "Gleam":utf8,
+            4.2:float,
+            <<
+              <<1, 2, 3>>:bit_string,
+              "Gleam":utf8,
+              1024
+            >>:bit_string
+          >>
+        "#
+    )
+}