+++ compiler-core/src/type_/environment.rs
@@ -23,6 +23,9 @@ pub struct Environment<'a> {
     pub unused_modules: HashMap<SmolStr, SrcSpan>,
     pub imported_types: HashSet<SmolStr>,
 
+    /// Names of modules that have been imported with as name.
+    pub imported_module_aliases: HashSet<SmolStr>,
+
     /// Values defined in the current function (or the prelude)
     pub scope: im::HashMap<SmolStr, ValueConstructor>,
 
@@ -73,6 +76,7 @@ impl<'a> Environment<'a> {
             scope: prelude.values.clone().into(),
             importable_modules,
             imported_types: HashSet::new(),
+            imported_module_aliases: HashSet::new(),
             current_module,
             warnings,
             entity_usages: vec![HashMap::new()],
@@ -131,7 +135,7 @@ impl<'a> Environment<'a> {
 
         // We only check for unused entities if the scope was successfully
         // processed. If it was not then any seemingly unused entities may have
-        // been used beyond the point where the error occured, so we don't want
+        // been used beyond the point where the error occurred, so we don't want
         // to incorrectly warn about them.
         if was_successful {
             self.handle_unused(unused);
@@ -384,7 +388,7 @@ impl<'a> Environment<'a> {
         hydrator: &Hydrator,
     ) -> Arc<Type> {
         match t.deref() {
-            Type::App {
+            Type::Named {
                 public,
                 name,
                 module,
@@ -394,7 +398,7 @@ impl<'a> Environment<'a> {
                     .iter()
                     .map(|t| self.instantiate(t.clone(), ids, hydrator))
                     .collect();
-                Arc::new(Type::App {
+                Arc::new(Type::Named {
                     public: *public,
                     name: name.clone(),
                     module: module.clone(),
@@ -448,7 +452,7 @@ impl<'a> Environment<'a> {
         match self
             .entity_usages
             .last_mut()
-            .expect("Attempted to access non-existant entity usages scope")
+            .expect("Attempted to access non-existent entity usages scope")
             .insert(name.clone(), (kind, location, false))
         {
             // Private types can be shadowed by a constructor with the same name
@@ -500,8 +504,12 @@ impl<'a> Environment<'a> {
         self.handle_unused(unused);
 
         for (name, location) in self.unused_modules.clone().into_iter() {
-            self.warnings
-                .emit(Warning::UnusedImportedModule { name, location });
+            let warning = if self.imported_module_aliases.contains(&name) {
+                Warning::UnusedImportedModuleAlias { name, location }
+            } else {
+                Warning::UnusedImportedModule { name, location }
+            };
+            self.warnings.emit(warning);
         }
     }
 
@@ -559,7 +567,7 @@ impl<'a> Environment<'a> {
         value_typ: Arc<Type>,
     ) -> Result<(), Vec<SmolStr>> {
         match &*value_typ {
-            Type::App {
+            Type::Named {
                 name: type_name,
                 module: module_name,
                 ..
@@ -679,13 +687,13 @@ pub fn unify(t1: Arc<Type>, t2: Arc<Type>) -> Result<(), UnifyError> {
 
     match (t1.deref(), t2.deref()) {
         (
-            Type::App {
+            Type::Named {
                 module: m1,
                 name: n1,
                 args: args1,
                 ..
             },
-            Type::App {
+            Type::Named {
                 module: m2,
                 name: n2,
                 args: args2,